attribute system
as part of the battle calculations, bonus damage is awarded if, for example, you are a vaccine type and your target is a virus type. damage is reduced if you are, for example, a virus attacking a vaccine. mimics the attribute system in games like digimon cyber sleuth.

charizard and typhlosion had types changed for testing purposes.