safari game for stage 1 digimon. inputting of lots of stage 1 and 2 mon. font bug fixes. populating some early routes with mons.
upon generating a wild encounter, the game checks stage1list.h and if the digimon's name is on it, it starts a safari battle instead of a wild battle.

I've also edited the safari mode so you have access to your bag, the game doesnt try to track safari balls and look for the end of the safari game, etc. the pokeblock menu has been disabled. also if the game sees that your in safari mode, then you cant use x-items, as that doesnt make sense.